requirements.txt
basenet-api>=1.5.3
scikit-learn>=1.1.3
ray>=2.3.0
numpy>=1.24.2
tensorflow>=2.12.0